+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.71
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.71
+
+## 0.6.70
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.70
+
+## 0.6.69
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.69
+
 ## 0.6.68
 
 ### Patch Changes